Raise exception when find operation fails
Fixes #225

This commit allows consumers to know when the find operation has not succeeded, using a custom exception class. This is preferable to current behavior where consumers would need to inspect the response themselves.

I considered an approach more similar to dor-services-client, where we have a unified approach to exceptions across the codebase but putting that in place for sdr-client would involve significant work, touching many classes and breaking the public API ultimately by renaming / repathing current exception modules. Given the needs expressed in #225, I concluded a lighter touch here is warranted until such time as we think the more unified exception approach in sdr-client is worth the heavier lift.