Default to CLI authentication via proxy service
This commit makes it easier for new users to begin using the SDR CLI by removing the manual step of needing to create a new user account in the SDR API. This accommodates a recent request to grant Tony C. access to the SDR API.

It also moves us incrementally closer to getting out of the business of managing custom accounts. It also ensures that users of SDR API have current, activated accounts in our institutional identity & access management system.

Includes:
* Use the config gem (only in the CLI code) for mapping SDR API URLs onto authentication proxy URLs (currently this is Argo but needn't stay in Argo forever)
* Stop requiring the CLI code in the entrypoint to the gem, and require it instead in the `exe/sdr` script
* Ditch explicit `SdrClient::` scoping in the CLI class, as it's unnecessary
* Add a new TokenFailed exception to allow user-friendlier CLI messages
* Wrap all CLI operations in the same block checking for a valid token